Added a compose() function to compose promises.
The compose() function takes multiple promises and composes them into a
single promise that will be fulfilled when the rest of the promises are
fulfilled, and will fail immediately if any of the given promises fail.

It passes the first error message it gets as the errback's failure
message, and it passes an in-order array of data it receives to the
callback.